+++ Sources/ParseSwift/Objects/ParseRole.swift
@@ -104,13 +104,6 @@ public extension ParseRole {
         self.ACL = acl
     }
 
-    func hash(into hasher: inout Hasher) {
-        let name = self.name ?? self.objectId
-        hasher.combine(name)
-        hasher.combine(createdAt)
-        hasher.combine(updatedAt)
-    }
-
     func mergeParse(with object: Self) throws -> Self {
         guard hasSameObjectId(as: object) else {
             throw ParseError(code: .otherCause,